backend/requirements.txt
fastapi
google-api-python-client
google-auth
google-auth-oauthlib
google-genai
langchain
langgraph
streamlit
uvicorn
google-generativeai
dotenv
google-auth-httplib2
supabase
dateparser